% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefiles-docs.R
\docType{data}
\name{arctic_bathy}
\alias{arctic_bathy}
\title{Arctic Stereographic bathymetry}
\format{
\code{\link[sp:SpatialPolygons]{SpatialPolygonsDataFrame}} bathymetry in Arctic Stereographic projection (EPSG:3995). Contains latitudes between 30 and 90 degrees. eneralized from ETOPO1 grid.
}
\source{
Amante, C. and B.W. Eakins, 2009. ETOPO1 1 Arc-Minute Global Relief Model: Procedures, Data Sources and Analysis. NOAA Technical Memorandum NESDIS NGDC-24. National Geophysical Data Center, NOAA. \doi{10.7289/V5C8276M}.
}
\usage{
arctic_bathy
}
\description{
Arctic Stereographic bathymetry
}
\seealso{
Other mapfiles: 
\code{\link{antarctic_bathy}},
\code{\link{antarctic_glacier}},
\code{\link{antarctic_land}},
\code{\link{arctic_glacier}},
\code{\link{arctic_land}},
\code{\link{bs_bathy}},
\code{\link{bs_glacier}},
\code{\link{bs_land}},
\code{\link{dd_bathy}},
\code{\link{dd_glacier}},
\code{\link{dd_land}}
}
\concept{mapfiles}
\keyword{maps}
\keyword{shapefiles}
