% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefiles-docs.R
\docType{data}
\name{arctic_land}
\alias{arctic_land}
\title{Arctic Stereographic land shapes}
\format{
\link[sf:st_sf]{Simple feature collection} land shapes in Arctic Stereographic projection (EPSG:3995). Contains latitudes between 30 and 90 degrees. Obtained from Natural Earth Data (10m vectors). Includes the islands dataset.
}
\source{
\href{https://www.naturalearthdata.com/}{Natural Earth Data}
}
\usage{
arctic_land
}
\description{
Arctic Stereographic land shapes
}
\seealso{
Other mapfiles: 
\code{\link{antarctic_bathy}},
\code{\link{antarctic_glacier}},
\code{\link{antarctic_land}},
\code{\link{arctic_bathy}},
\code{\link{arctic_glacier}},
\code{\link{bs_bathy}},
\code{\link{bs_glacier}},
\code{\link{bs_land}},
\code{\link{dd_bathy}},
\code{\link{dd_glacier}},
\code{\link{dd_land}}
}
\concept{mapfiles}
\keyword{maps}
\keyword{shapefiles}
