% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefiles-docs.R
\docType{data}
\name{dd_bathy}
\alias{dd_bathy}
\title{Decimal degree bathymetry}
\format{
\code{\link[sp:SpatialPolygons]{SpatialPolygonsDataFrame}} bathymetry in decimal degrees (+init=epsg:4326). Contains latitudes between -70 and 70 degrees. Generalized from ETOPO1 grid.
}
\source{
Amante, C. and B.W. Eakins, 2009. ETOPO1 1 Arc-Minute Global Relief Model: Procedures, Data Sources and Analysis. NOAA Technical Memorandum NESDIS NGDC-24. National Geophysical Data Center, NOAA. \doi{10.7289/V5C8276M}
}
\usage{
dd_bathy
}
\description{
Decimal degree bathymetry
}
\seealso{
Other shapefiles: 
\code{\link{antarctic_bathy}},
\code{\link{antarctic_glacier}},
\code{\link{antarctic_land}},
\code{\link{arctic_bathy}},
\code{\link{arctic_glacier}},
\code{\link{arctic_land}},
\code{\link{bs_bathy}},
\code{\link{bs_glacier}},
\code{\link{bs_land}},
\code{\link{dd_glacier}},
\code{\link{dd_land}},
\code{\link{fishingAreasNor}}
}
\concept{shapefiles}
\keyword{datasets}
\keyword{shapefiles}
