% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefiles-docs.R
\docType{data}
\name{dd_glacier}
\alias{dd_glacier}
\title{Decimal degree glacier shapes}
\format{
\code{\link[sp:SpatialPolygons]{SpatialPolygonsDataFrame}} glacier shapes in decimal degrees (+init=epsg:4326). Contains latitudes between -70 and 70 degrees. Obtained from Natural Earth Data (10m vectors). Includes the ice-sheets dataset.
}
\source{
\href{https://www.naturalearthdata.com/}{Natural Earth Data}
}
\usage{
dd_glacier
}
\description{
Decimal degree glacier shapes
}
\seealso{
Other mapfiles: 
\code{\link{antarctic_bathy}},
\code{\link{antarctic_glacier}},
\code{\link{antarctic_land}},
\code{\link{arctic_bathy}},
\code{\link{arctic_glacier}},
\code{\link{arctic_land}},
\code{\link{bs_bathy}},
\code{\link{bs_glacier}},
\code{\link{bs_land}},
\code{\link{dd_bathy}},
\code{\link{dd_land}}
}
\concept{mapfiles}
\keyword{maps}
\keyword{shapefiles}
