% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefiles-docs.R
\docType{data}
\name{fishingAreasNor}
\alias{fishingAreasNor}
\title{Major fishing zones of Norway}
\format{
\code{\link[sf:st_sfc]{sf object}} in decimal degrees (EPSG:4326) containing major fishing zones defined by the Norwegian Directorate of Fisheries.
}
\source{
\href{https://kart.fiskeridir.no/stat}{Norwegian Directorate of Fisheries}
}
\usage{
fishingAreasNor
}
\description{
Major fishing zones of Norway
}
\seealso{
Other datasets: 
\code{\link{ices_areas}}
}
\concept{datasets}
\keyword{datasets}
\keyword{fishery}
\keyword{shapefiles}
