% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggOceanMapsData.R
\docType{package}
\name{ggOceanMapsData-package}
\alias{ggOceanMapsData}
\alias{ggOceanMapsData-package}
\title{Spatial Data Repository for ggOceanMaps}
\description{
This is a spatial data repository package for the \pkg{ggOceanMaps} package.
The package provides low-resolution shapefiles required by \pkg{ggOceanMaps}. 
These shapefiles are placed in a separate package due to CRAN Repository 
Policy and package size limits. Please see the 
\href{https://github.com/MikkoVihtakari/ggOceanMaps}{ggOceanMaps} 
documentation for further information.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mikkovihtakari.github.io/ggOceanMaps}
  \item \url{https://github.com/MikkoVihtakari/ggOceanMapsData}
  \item Report bugs at \url{https://github.com/MikkoVihtakari/ggOceanMaps/issues}
}

}
\author{
\strong{Maintainer}: Mikko Vihtakari \email{mikko.vihtakari@hi.no} (\href{https://orcid.org/0000-0003-0371-4319}{ORCID}) (Institute of Marine Research)

}
\keyword{internal}
