% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefiles-docs.R
\docType{data}
\name{ices_areas}
\alias{ices_areas}
\title{ICES Advisory Areas}
\format{
\code{\link[sf:st_sfc]{sf object}} in decimal degrees (EPSG:4326) containing ICES Advisory Areas
}
\source{
\href{https://www.ices.dk/}{International Council for the Exploration of the Sea}
}
\usage{
ices_areas
}
\description{
ICES Advisory Areas
}
\seealso{
Other datasets: 
\code{\link{fishingAreasNor}}
}
\concept{datasets}
\keyword{datasets}
\keyword{fishery}
\keyword{shapefiles}
