% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefiles-docs.R
\docType{data}
\name{antarctic_bathy}
\alias{antarctic_bathy}
\title{Antarctic Stereographic bathymetry}
\format{
\link[sf:st_sf]{Simple feature collection} bathymetry in Antarctic Stereographic projection (EPSG:3031). Contains latitudes between -80 and -20 degrees. Generalized from ETOPO1 grid.
}
\source{
Amante, C. and B.W. Eakins, 2009. ETOPO1 1 Arc-Minute Global Relief Model: Procedures, Data Sources and Analysis. NOAA Technical Memorandum NESDIS NGDC-24. National Geophysical Data Center, NOAA. \doi{10.7289/V5C8276M}
}
\usage{
antarctic_bathy
}
\description{
Antarctic Stereographic bathymetry
}
\seealso{
Other mapfiles: 
\code{\link{antarctic_glacier}},
\code{\link{antarctic_land}},
\code{\link{arctic_bathy}},
\code{\link{arctic_glacier}},
\code{\link{arctic_land}},
\code{\link{bs_bathy}},
\code{\link{bs_glacier}},
\code{\link{bs_land}},
\code{\link{dd_bathy}},
\code{\link{dd_glacier}},
\code{\link{dd_land}}
}
\concept{mapfiles}
\keyword{maps}
\keyword{shapefiles}
